Refactoring: merge video-watch scope into common.
In order to be able to use lit in videowatch related code, we have to
merge these scopes.
